modular SubmissionList and split it from SubmissionView
